refactored generateBreaks tests
